Add test for ceq framework
