+++ setup.py
@@ -7,7 +7,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='OpenFisca-COTE_D_IVOIRE',
-    version='0.8.0',
+    version='0.9.0',
     author='OpenFisca Team',
     author_email='contact@openfisca.fr',
     description=u'OpenFisca tax and benefit system for COTE_D_IVOIRE',
@@ -27,8 +27,12 @@ setup(
             "pytest <5.0",
             "scipy >= 0.17",  # Only used to test de_net_a_brut reform
             "requests >= 2.8",
+            "openfisca-ceq",
             "openfisca-survey-manager[dev] >= 0.17.4",
             "yamllint >=1.11.1,<1.15",
+            ],
+        'ceq': [
+            "openfisca-ceq >= 0.2.3",
             ],
         },
     packages=find_packages(),
